feat: add category list as component, edit styles
