+++ src/entities/user/index.tsx
@@ -1,2 +1,4 @@
-const temp = (): void => {};
-export default temp;
+import UserProfile from './ui/userProfile/UserProfile';
+import UserAddress from './ui/userAddress/UserAddress';
+
+export { UserProfile, UserAddress };
