+++ src/shared/categories/model/categoriesSelectors.ts
@@ -1,5 +1,5 @@
-import { RootState } from '../../app/appStore/appStore';
-import { CatalogCategory } from '../types/types';
+import { RootState } from '../../../app/appStore/appStore';
+import { CatalogCategory } from '../../types/types';
 
 const selectCategories = (state: RootState): CatalogCategory[] => state.categoriesData.categories;
