+++ src/shared/categories/ui/CategoriesList.module.scss
@@ -1,0 +1,74 @@
+@import "../../../app/styles/variables.scss";
+
+.categoriesItem {
+  display: flex;
+  flex-direction: column;
+  background: transparent;
+  text-align: start;
+  padding: 4px 0px;
+
+  .link {
+    width: 100%;
+    display: flex;
+  }
+}
+
+.subCategoriesList {
+  display: flex;
+  flex-wrap: wrap;
+  flex-direction: column;
+  max-height: 280px;
+  position: absolute;
+  left: 250px;
+  top: 66px;
+
+  @media (max-width: $width-mobile-big) {
+    position: static;
+    max-height: 100%;
+  }
+
+  li {
+    padding: 20px;
+
+    @media (max-width: $width-mobile-big) {
+      padding: 20px 0 0 20px;
+    }
+  }
+
+  a {
+    color: var(--color-primary-text);
+    font-weight: var(--font-weight-regular);
+  }
+  a:hover {
+    cursor: pointer;
+    color: var(--color-hovered-btn);
+  }
+}
+
+.link {
+  display: flex;
+  gap: 20px;
+  color: var(--color-primary-text);
+  transition: ease-in-out 0.2s;
+  width: 100%;
+}
+
+.arrow {
+  height: 20px;
+  width: 20px;
+  background: none;
+  color: var(--color-categories-arrow-button);
+  transition: ease-in-out 0.2s;
+}
+
+.link:hover,
+.link:hover ~ .arrow {
+  cursor: pointer;
+  color: var(--color-hovered-btn);
+}
+
+.rotatedArrow {
+  @media (max-width: $width-mobile-big) {
+    transform: rotate(90deg);
+  }
+}
