+++ src/widgets/header/ui/nav/catalog/categories/categories.module.scss
@@ -30,7 +30,7 @@
   display: flex;
   flex-direction: column;
   gap: 20px;
-  max-width: 250px;
+  max-width: 300px;
   padding: 20px 30px;
   border-right: var(--color-categories-arrow-button) 2px solid;
 
@@ -48,38 +48,6 @@
   }
 }
 
-.subCategoriesList {
-  display: flex;
-  flex-wrap: wrap;
-  flex-direction: column;
-  max-height: 280px;
-  position: absolute;
-  left: 250px;
-  top: 66px;
-
-  @media (max-width: $width-mobile-big) {
-    position: static;
-    max-height: 100%;
-  }
-
-  li {
-    padding: 20px;
-
-    @media (max-width: $width-mobile-big) {
-      padding: 20px 0 0 20px;
-    }
-  }
-
-  a {
-    color: var(--color-primary-text);
-    font-weight: var(--font-weight-regular);
-  }
-  a:hover {
-    cursor: pointer;
-    color: var(--color-hovered-btn);
-  }
-}
-
 .link {
   color: var(--color-primary-text);
   transition: ease-in-out 0.2s;
@@ -88,28 +56,6 @@
   &:hover {
     cursor: pointer;
     color: var(--color-hovered-btn);
-  }
-}
-
-.arrow {
-  margin-left: 10px;
-  height: 20px;
-  width: 20px;
-  background: none;
-  padding: 4px;
-  color: var(--color-categories-arrow-button);
-  transition: ease-in-out 0.2s;
-
-  &:hover {
-    cursor: pointer;
-    color: var(--color-hovered-btn);
-  }
-}
-
-.rotatedArrow {
-  @media (max-width: $width-mobile-big) {
-    transform: rotate(90deg);
-    padding: 0;
   }
 }
