+++ src/widgets/header/ui/nav/index.tsx
@@ -8,7 +8,7 @@ import Catalog from './catalog';
 
 import RoutesName from '../../../../shared/routing';
 
-import getCategoriesAction from '../../../../shared/categories/categoriesAction';
+import getCategoriesAction from '../../../../shared/categories/model/categoriesAction';
 import { useAppDispatch } from '../../../../app/appStore/hooks';
 
 function HeaderNav(): JSX.Element {
